rf-#17: complete profile/userInputData page
