+++ src/App/App.tsx
@@ -33,7 +33,6 @@ const {UserContextValue,userLoginContextValue, isModalOpen} =  useIsUserRegister
         <div className="loader-container">
         <span className="loader"></span>
         </div>
-
         </Modal>}
     <Outlet />
     </div>
