+++ src/components/Card/CardWrapper.ts
@@ -1,6 +1,7 @@
 import styled from "styled-components";
+import { URL } from "../../constants/endpoints";
 
-const CardWrapper = styled.div`
+const CardWrapper = styled.div<{id?: string}>`
 width: 60%;
 flex-shrink: 0;
 margin: 5px;
@@ -14,7 +15,8 @@ margin: 5px;
         width: 100%;
         height: 60%;
         border-radius: 10px;
-        background-image: url("https://i.pinimg.com/564x/bd/78/30/bd7830482a990a97c57300a6f6a4bade.jpg");
+    
+        background-image: url(${props=> props.id ? `${URL().endpoint}/upload/${props.id}`:"https://i.pinimg.com/564x/bd/78/30/bd7830482a990a97c57300a6f6a4bade.jpg"});
         background-size: cover;
     }
