+++ src/components/HourSelect/type.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react"
+
+type THourProp = {
+    disabled?: boolean,
+    userClosingHour?: number,
+    userOpeningHour?:number,
+    children?: ReactNode,
+    onTimeChange: (e: React.ChangeEvent<HTMLSelectElement>) => void
+}
+
+export type {THourProp}
