+++ src/components/LocationSelect/index.tsx
@@ -1,8 +1,8 @@
-import { GoLocation, GoPencil } from "react-icons/go"
+import { GoLocation } from "react-icons/go"
 import { colors } from "../../constants/colors"
 import useGetStates from "../LocationNav/hooks/useGetState"
 import { TSelect } from "./type"
-import styled from "styled-components"
+import LocationSelectWrapper from "./LocationSelectWrapper"
 
 
 
@@ -29,29 +29,3 @@ const LocationSelect = ({handleSelect, disabled, value, children}:TSelect)=>{
 export default LocationSelect
 
 
-const LocationSelectWrapper = styled.div`
-width: 100%;
-
-     .box {
-                    width: 100%;
-                    height: 60px;
-                    border: 1px solid red;
-                    display: flex;
-                    justify-content: space-between;
-                    align-items: center;
-                    margin-top: 8px;
-                    border: 2px solid ${colors.primary.lightGray};
-                    border-radius: 12px;
-                    padding: 12px;
-                    margin-bottom: 30px;
-
-                    select {
-                    width: 95%;
-                    height: 40px;
-                    background-color: ${colors.primary.white};
-                    border: none;
-                    padding: 8px;
-                    outline: none;
-                  }
-                  }
-`
