+++ src/components/Verification/index.tsx
@@ -2,10 +2,11 @@ import { GoClock,GoStar, GoVerified } from "react-icons/go";
 import VerificationWrapper from "./VerificationWrapper";
 import { colors } from "../../constants/colors";
 import { TVerification } from "./type";
+import getTime from "../../utils/getTime";
 
 
 
-const ProfileDetailsVerification = ({isVerified,rating}: TVerification) => {
+const ProfileDetailsVerification = ({isVerified,rating,time}: TVerification) => {
   return (
     <VerificationWrapper>
       <div className="verify-container">
@@ -19,7 +20,7 @@ const ProfileDetailsVerification = ({isVerified,rating}: TVerification) => {
           <div className="icons-wrapper">
             <GoClock  color={colors.primary.cyan}/>
           </div>
-          <p>7PM - 10AM</p>
+          <p>{getTime(+time.from)} - {getTime(+time.to)}</p>
         </div>
         <div className="verify-rate">
         <GoVerified  color={colors.primary.darkGreen}/>
