+++ src/pages/ChildcareProfile/index.tsx
@@ -1,153 +1,51 @@
+import { GoChevronLeft,GoKebabHorizontal, GoPencil, GoUpload } from "react-icons/go";
 import UserInputDataWrapper from "./UserInputDataWrapper";
 import NameInput from "../../components/NameInput";
-import { GoChevronLeft,GoKebabHorizontal, GoPencil, GoUpload } from "react-icons/go";
 import { colors } from "../../constants/colors";
 import Header from "../../components/Header";
-import { useEffect,  useRef, useState, useTransition } from "react";
-import useUploadImage from "./hooks/useUploadImage";
 import RadioInput from "../../components/RadioInput";
 import LocationSelect from "../../components/LocationSelect";
-import { TLoginResponse } from "../Auth/Login/type";
-import useGetProfile from "./hooks/useGetProfile";
 import HourSelect from "../../components/HourSelect";
-import { useSearchParams } from "react-router-dom";
-import {TChild, TExtendChildcare} from "./type";
-import useSaveProfile from "./hooks/useSaveProfile";
-import ReactDOM from "react-dom"
-import _ from "lodash" 
-import { start } from "repl";
+import Modal from "../../components/Modal";
+import useUploadEffect from "./hooks/useUploadEffect";
 
 
 
 
 const ChildCareProfile = () => {
-
-  const ref = useRef<HTMLInputElement | null>(null)
-  const {response, mutateUpload,error,isLoading} = useUploadImage()
-  const [id,setId] = useState<string>()
-  const radioRef = useRef<HTMLDivElement | null>(null) 
-  const [isDaycareOwner, setIsDaycareOwner] =  useState<boolean>()
-  const {setId:setUserId, data} = useGetProfile()
-  const [label, setLabel] = useSearchParams()
-  const {mutateProfile, response:profileResponse} = useSaveProfile()
-
-  const [form, setForm] = useState<TExtendChildcare<TChild & {location:string},"_id">>({
-    amount: "",
-    description: "",
-    exactLocation: "",
-    image: "",
-    isOpen: false,
-    location: "",
-    owner: "",
-    phonenumber: "",
-    rating: 3,
-    title: "",
-    userId: "",
-    from: "",
-    to: "",
-    role:""
-  })
-
+const {form,
+  handleMimicUpload,
+  handleSelect,
+  isDaycareOwner,
+  isProfileLoading,
+  isUpdateLoading,
+  onAmountChanged,
+  onClosingTimeChange,
+  onDescriptionChange,
+  onFileUploaded,
+  onMobileNumberChange,
+  onOpeningTimeChange,
+  onRoleChange,
+  onSubmit,
+  onTitleChange,
+  radioRef,
+  setForm,
+  isLoading,
+  label,
+  setLabel,
+  error,
+  ref,
+  errorResponse
+} = useUploadEffect()
   
-
-  const handleMimicUpload = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>)=>{
-    e.preventDefault()
-    if(ref && ref.current){
-      ref.current.click()
-    }
-  }
-
-  const onFileUploaded = (e: React.ChangeEvent<HTMLInputElement>)=>{
-    e.preventDefault()
-    if(e.target.files){
-    mutateUpload(e.target.files[0])
-    }
-  }
-
-  const handleSelect = (e: React.ChangeEvent<HTMLSelectElement>)=>{
- 
-    setForm((prev)=> ({...prev, location: e.target.value}))
-  }
-
-
-  useEffect(()=>{
-    if(response){
-      setId(response.id)
-    }
-  },[response])
-
-  useEffect(()=>{
-    let user = localStorage.getItem("DayCareuserLoginInfo")
-    if(user){
-      let parse = JSON.parse(user) as TLoginResponse
-      setIsDaycareOwner(parse.message.day_care_owner)
-      setUserId(parse.message._id)
-    }
-  },[])
-
-  useEffect(()=>{
-if(data){
-  let currentData = {..._.omit(data,["_id","isVerified","__v","perDuration"]),location: data.exactLocation, amount: data.amount.toString()}
-    setForm(currentData)
-
-  setId(form.image)
-}
-  }, [data])
-
-    // useEffect(()=>{
-    //   if(profileResponse){
-    
-    //     let currentData = {..._.omit(profileResponse,["_id","isVerified","__v"]),location: profileResponse.exactLocation, amount: profileResponse.amount}
-    //     setForm(currentData)
-    //     setId(form.image)
-    //   }
-    //     }, [profileResponse])
-  
-  const onTitleChange = (e: React.ChangeEvent<HTMLInputElement>)=>{
-    e.preventDefault()
-     setForm((prev)=> ({...prev, title: e.target.value}))
-  }
-
-  const onMobileNumberChange = (e: React.ChangeEvent<HTMLInputElement>)=>{
-    e.preventDefault()
-    setForm((prev)=> ({...prev, phonenumber: e.target.value}))
-  }
-
-  const onDescriptionChange = (e: React.ChangeEvent<HTMLTextAreaElement>)=> {
-    e.preventDefault()
-    setForm((prev)=> ({...prev, description: e.target.value}))
-  }
-
-  const onAmountChanged = (e: React.ChangeEvent<HTMLInputElement>)=>{
-e.preventDefault()
-setForm((prev)=> ({...prev, amount: e.target.value}))
-  }
-
-  const onOpeningTimeChange = (e: React.ChangeEvent<HTMLSelectElement>)=>{
-    e.preventDefault()
-    setForm((prev)=> ({...prev, from: e.target.value}))
-  }
-  const onClosingTimeChange = (e: React.ChangeEvent<HTMLSelectElement>)=>{
-    e.preventDefault()
-    setForm((prev)=> ({...prev, to: e.target.value}))
-  }
-
-
-  const onSubmit = (e: React.FormEvent<HTMLFormElement>) =>{
-    e.preventDefault()
-  
-    let sanitizedData = _.omitBy(form, (value)=> value === "") as Partial<TExtendChildcare<TChild & {location:string},"_id">>
-    mutateProfile(_.omit(sanitizedData, ["exactLocation"]))
-  }
-
-  const onRoleChange = (e: React.ChangeEvent<HTMLInputElement>)=>{
-    e.preventDefault()
-    setForm((prev)=> ({...prev, role: e.target.value}))
-  }
-  console.log(form)
 
   return (
-    <UserInputDataWrapper id={id}>
+    <UserInputDataWrapper id={form.image}>
+      {isProfileLoading ? <Modal ismodalopen={JSON.stringify(isProfileLoading)}>
+      <div className="loader-container">
+        <span className="loader"></span>
+        </div>
+      </Modal>:""}
       <div className="profile-wrapper">
        <Header title="Profile">
         <GoChevronLeft />
@@ -291,9 +189,12 @@ setForm((prev)=> ({...prev, amount: e.target.value}))
           </HourSelect>
           </div>
         </div>
-
-        <button type="submit">save</button>
+        {errorResponse ?  <div>{errorResponse.message}</div> : ""}
         </form>
+
+        <div className="btnwrapper">
+            <button type="submit" onClick={(e)=> onSubmit(e)} className="save-btn">{isUpdateLoading ? "saving...":  "save"}</button>
+            </div>
       </div>
     </UserInputDataWrapper>
   );
