+++ src/pages/ChildcareProfile/type.ts
@@ -3,4 +3,9 @@ import { TChildCare } from "../Home/type"
 type TChild = Omit<TChildCare, "location" | "isVerified" | "perDuration" >
 type TExtendChildcare<T, K extends keyof  T> = Omit<T, K>
 
-export type {TExtendChildcare, TChild} 
+type TMessage = {
+    message: string,
+    id: string
+}
+
+export type {TExtendChildcare, TChild, TMessage}
